Cleaned up enemy.rs and move it to enemy_util.rs where all resourses and componenets live.
